% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{read.wordvectors}
\alias{read.wordvectors}
\title{Read word vectors from a word2vec model from disk}
\usage{
read.wordvectors(file, normalize = TRUE, n = .Machine$integer.max)
}
\arguments{
\item{file}{the path to the model file}

\item{normalize}{logical indicating to normalize the embeddings by dividing by the factor (sqrt(sum(x . x) / length(x))). Defaults to TRUE.}

\item{n}{integer, indicating to limit the number of words to read in. Defaults to reading all words.}
}
\value{
a matrix with the embeddings of the words, which are in UTF-8 encoding
}
\description{
Read word vectors from a word2vec model from disk
}
\examples{
path  <- system.file(package = "word2vec", "models", "example.bin")
embed <- read.wordvectors(path, n = 10)
embed <- read.wordvectors(path)
}
